python/requirements.txt
tensorflow>= 2.3
opencv-python
bidict == 0.21.2
websocket-client==0.57.0
python-socketio == 5.0.4
python-engineio == 4.0.0
